+++ components/41-organisms/teaser-collection/teaser-collection.binding.js
@@ -1,0 +1,13 @@
+/**
+ * @file
+ */
+(function ($) {
+  'use strict'
+    $(window).on('load', function (e) {
+        $('.js-masonry-layout').masonry({
+          itemSelector: 'li',
+          columnWidth: 'li',
+          percentPosition: true
+        });
+    });
+})(jQuery);